Classification and Clustering algorithms
In a given data-set, 70% of the data are taken as TrainData and the rest
30% is considered as TestData. For Supervised algorithms, Each
classifier is trained with the TrainData, and then the labels are
predicted in the test data, along with the accuracy of the resultant
classification.
Further, few transformations like PCA, LDA and feature selection are
implemented to reduce the data size to be classified, thereby grouping
similar data, and reducing the run time along with improved
classification accuracy.
The Algorithms implemented include KNN, SVM, Centroid method, Linear
Regression and Kmeans clustering.